Fixed retryDialog button ordering. Fixed default action on confirmation dialog. Set notifications to be silent. AddToQueue button implemented on channel fragment. Added confirmation for deleting downloaded playlist videos. Separately handled UnknownHostException to prevent the dialog from showing when offline.

+++ app/src/main/java/com/futo/platformplayer/fragment/mainactivity/main/PlaylistFragment.kt
@@ -284,13 +284,17 @@ class PlaylistFragment : MainFragment() {
                 _buttonDownload.setImageResource(R.drawable.ic_loader_animated);
                 _buttonDownload.drawable.assume<Animatable, Unit> { it.start() };
                 _buttonDownload.setOnClickListener {
-                    StateDownloads.instance.deleteCachedPlaylist(playlist.id);
+                    UIDialogs.showConfirmationDialog(context, "Are you sure you want to delete the downloaded videos?", {
+                        StateDownloads.instance.deleteCachedPlaylist(playlist.id);
+                    });
                 }
             }
             else if(isDownloaded) {
                 _buttonDownload.setImageResource(R.drawable.ic_download_off);
                 _buttonDownload.setOnClickListener {
-                    StateDownloads.instance.deleteCachedPlaylist(playlist.id);
+                    UIDialogs.showConfirmationDialog(context, "Are you sure you want to delete the downloaded videos?", {
+                        StateDownloads.instance.deleteCachedPlaylist(playlist.id);
+                    });
                 }
             }
             else {
